# 5Gr-SO/nbi/python-flask-server/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST-API server for the North Bound Interface of the 5Growth-SO. Find\
    \ more at http://5growth.eu"
  version: "1.0"
  title: "5Growth-SO NBI"
host: "localhost:8080"
basePath: "/5gt/so/v1"
...
schemes:
- "http"
paths:
  /ns:
    post:
      tags:
      - "network service"
      summary: "Creates and returns a Network Service identifier (nsId)"
      description: ""
      operationId: "create_ns_identifier"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Network Service information"
        required: true
        schema:
          $ref: "#/definitions/CreateNsIdentifierRequest"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/inline_response_201"
        400:
          description: "Bad Request"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/{nsId}:
    get:
      tags:
      - "network service"
      summary: "Returns information of the network service referenced by nsId"
      description: ""
      operationId: "query_ns"
      produces:
      - "application/json"
      parameters:
      - name: "nsId"
        in: "path"
        description: "ID of the network service"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/NsInfo"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/{nsId}/instantiate:
    put:
      tags:
      - "network service"
      summary: "Instantiates the Network Service referenced by nsId"
      description: ""
      operationId: "instantiate_ns"
      produces:
      - "application/json"
      parameters:
      - name: "nsId"
        in: "path"
        description: ""
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Network Service information"
        required: true
        schema:
          $ref: "#/definitions/NsInstantiationRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/inline_response_200"
        400:
          description: "Bad Request"
        404:
          description: "Not found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/{nsId}/scale:
    put:
      tags:
      - "network service"
      summary: "Scales the Network Service referenced by nsId"
      description: ""
      operationId: "scale_ns"
      produces:
      - "application/json"
      parameters:
      - name: "nsId"
        in: "path"
        description: "Identifier of the NS to be scaled"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Scale information"
        required: true
        schema:
          $ref: "#/definitions/NsScaleRequest"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/inline_response_200"
        400:
          description: "Bad Request"
        404:
          description: "Not found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
...
  /ns/{nsId}/terminate:
    put:
      tags:
      - "network service"
      summary: "Terminates the Network Service identified by nsId."
      description: ""
      operationId: "terminate_ns"
      produces:
      - "application/json"
      parameters:
      - name: "nsId"
        in: "path"
        description: "ID of the network service"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/inline_response_200"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/nsd:
    get:
      tags:
      - "network service"
      summary: "Returns information of all NSDs onboarded in the SO"
      description: ""
      operationId: "query_nsds"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/nsd/{nsdId}/{version}:
    get:
      tags:
      - "network service"
      summary: "Returns information of the network service referenced by nsdId"
      description: ""
      operationId: "query_nsd"
      produces:
      - "application/json"
      parameters:
      - name: "nsdId"
        in: "path"
        description: "ID of the network service descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the network service descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
    delete:
      tags:
      - "network service"
      summary: "Delete the onboarded network service referenced by nsdId"
      description: ""
      operationId: "delete_nsd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "nsdId"
        in: "path"
        description: "ID of the network service descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the network service descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/inline_response_201"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/vnfd:
    get:
      tags:
      - "network service"
      summary: "Returns information of all VNFDs onboarded in the SO"
      description: ""
      operationId: "query_vnfds"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/vnfd/{vnfdId}/{version}:
    get:
      tags:
      - "network service"
      summary: "Returns information of the virtual network function referenced by\
        \ vnfId"
      description: ""
      operationId: "query_vnfd"
      produces:
      - "application/json"
      parameters:
      - name: "vnfdId"
        in: "path"
        description: "ID of the virtual network function descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the virtual network function descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
    delete:
      tags:
      - "network service"
      summary: "Deletes the specified virtual network function"
      description: ""
      operationId: "delete_vnfd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "vnfdId"
        in: "path"
        description: "ID of the virtual network function descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the virtual network function descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/appd/{appdId}/{version}:
    get:
      tags:
      - "network service"
      summary: "Returns information of the MEC app function referenced by appdId"
      description: ""
      operationId: "query_appd"
      produces:
      - "application/json"
      parameters:
      - name: "appdId"
        in: "path"
        description: "ID of the MEC app descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the MEC app descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
    delete:
      tags:
      - "network service"
      summary: "Deletes the specified MEC application"
      description: ""
      operationId: "delete_appd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "appdId"
        in: "path"
        description: "ID of the MEC app descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the MEC app descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/pnfd/{pnfdId}/{version}:
    get:
      tags:
      - "network service"
      summary: "Returns information of the Physical Network Function (PNF) referenced\
        \ by pnfdId"
      description: ""
      operationId: "query_pnfd"
      produces:
      - "application/json"
      parameters:
      - name: "pnfdId"
        in: "path"
        description: "ID of the physical network function descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the physical network function descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
    delete:
      tags:
      - "network service"
      summary: "Delete the onboarded physical network function referenced by pnfdId"
      description: ""
      operationId: "delete_pnfd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "pnfdId"
        in: "path"
        description: "ID of the physical network function descriptor"
        required: true
        type: "string"
      - name: "version"
        in: "path"
        description: "Version of the physical network function descriptor"
        required: true
        type: "string"
      responses:
        200:
          description: "Successful Operation"
          schema:
            type: "object"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/vnfdManagement/vnfPackage:
    post:
      tags:
      - "network service"
      summary: "Returns information of the onboarded virtual network function"
      description: ""
      operationId: "onboard_vnfd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Information about the VNF descriptor"
        required: true
        schema:
          $ref: "#/definitions/VnfOnboardingRequest"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/VnfOnboardingReply"
        400:
          description: "Bad Request"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/appdManagement/appPackage:
    post:
      tags:
      - "network service"
      summary: "Returns information of the onboarded MEC application"
      description: ""
      operationId: "onboard_appd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Information about the MEC APP descriptor"
        required: true
        schema:
          $ref: "#/definitions/AppOnboardingRequest"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/AppOnboardingReply"
        400:
          description: "Bad Request"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/nsdManagement/nsd:
    post:
      tags:
      - "network service"
      summary: "Returns information of the onboarded network service"
      description: ""
      operationId: "onboard_nsd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The NSD descriptor"
        required: true
        schema:
          type: "object"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/NsOnboardingReply"
        400:
          description: "Bad Request"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
  /ns/pnfdManagement/pnfd:
    post:
      tags:
      - "network service"
      summary: "Returns information of the onboarded physical network function"
      description: ""
      operationId: "onboard_pnfd"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "The PNFD descriptor"
        required: true
        schema:
          type: "object"
      responses:
        200:
          description: "Successful Operation"
          schema:
            $ref: "#/definitions/PnfOnboardingReply"
        400:
          description: "Bad Request"
        404:
          description: "Not Found"
        500:
          description: "Internal Error"
      x-swagger-router-controller: "swagger_server.controllers.network_service_controller"
definitions:
  CreateNsIdentifierRequest:
    type: "object"
    required:
    - "nsName"
    - "nsdId"
    properties:
      nsdId:
        type: "string"
      nsName:
        type: "string"
      nsDescription:
        type: "string"
    xml:
      name: "NsIdentifierRequest"
    example:
      nsdId: "nsdId"
      nsName: "nsName"
      nsDescription: "nsDescription"
  NsInfo:
    type: "object"
    properties:
      status:
        type: "string"
      sapInfo:
        type: "array"
        items:
          $ref: "#/definitions/SapInfo"
    example:
      status: "status"
      sapInfo:
      - sapInstanceId: "sapInstanceId"
        sapName: "sapName"
        description: "description"
        address: "10.10.10.10"
      - sapInstanceId: "sapInstanceId2"
        sapName: "sapName2"
        description: "description2"
        address: "10.10.10.11"
  NsInstantiationRequest:
    type: "object"
    required:
    - "flavourId"
    - "nsInstantiationLevelId"
    properties:
      flavourId:
        type: "string"
      nsInstantiationLevelId:
        type: "string"
      serviceType:
        type: "string"
        enum:
        - "eMBB"
        - "uRLLC"
        - "mMTC"
      sapData:
        type: "array"
        items:
          $ref: "#/definitions/SapData_im"
      pnfInfo:
        type: "array"
        items:
          $ref: "#/definitions/PnfInfo_im"
      vnfInstanceData:
        type: "array"
        items:
          $ref: "#/definitions/VnfInstanceData_im"
      nestedNsInstanceId:
        type: "array"
        items:
          type: "string"
      locationConstraints:
        type: "array"
        items:
          $ref: "#/definitions/VnfLocationConstraint_im"
      additionalParamForNs:
        type: "object"
        additionalProperties:
          type: "string"
      additionalParamForVnf:
        type: "array"
        items:
          $ref: "#/definitions/ParamsForVnf_im"
      startTime:
        type: "string"
      additionalAffinityOrAntiAffinityRule:
        type: "array"
        items:
          $ref: "#/definitions/AffinityOrAntiAffinityRule_im"
    example:
      flavourId: "flavourId"
      nsInstantiationLevelId: "nsInstantiationLevelId"
  VnfOnboardingRequest:
    type: "object"
    required:
    - "checksum"
    - "name"
    - "provider"
    - "version"
    - "vnfPackagePath"
    properties:
      name:
        type: "string"
      version:
        type: "string"
      provider:
        type: "string"
      checksum:
        type: "string"
      vnfPackagePath:
        type: "string"
      userDefinedData:
        $ref: "#/definitions/userDefinedDatadef"
    example:
      name: "spr1"
      version: "0.2"
      provider: "ATOS"
      checksum: "VNFD_TEST_001_checksum"
      userDefinedData:
        pwd: "secret"
      vnfPackagePath: "http://{VS-IP}/CDN_SPR1_VNFD_0_2.tar"
...
  NsOnboardingReply:
    type: "object"
    required:
    - "nsdInfoId"
    properties:
      nsdInfoId:
        type: "string"
    example:
      nsdInfoId: "vCDN_v02"
  NsScaleRequest:
    type: "object"
    required:
    - "scaleType"
    properties:
      scaleType:
        type: "string"
        enum:
        - "SCALE_NS"
        - "SCALE_VNF"
      scaleNsData:
        $ref: "#/definitions/ScaleNsData_im"
      scaleVnfData:
        type: "array"
        items:
          $ref: "#/definitions/ScaleVnfData_im"
      scaleTime:
        type: "string"
...
  inline_response_201:
    required:
    - "nsId"
    properties:
      nsId:
        type: "string"
    example:
      nsId: "nsId"
  inline_response_200:
    required:
    - "operationId"
    properties:
      operationId:
        type: "string"
    example:
      operationId: "operationId"
...